fix/improve operators and related functions
There is a great boilerplate in the code of most operators
in this class. And repetition is bad for many reasons. One
way to remove repetition is by metaprogramming, like macros,
and templates.

I started with the unique functions mul and div, then...

I used C macros to generate:
- the element-wise member functions: add, sub, amul, adiv;
- the equiv. member operators: +=, -=, *=, /=,
  overloaded to deal with matrices and doubles;
- the equiv. nonmember operators: +, -, *, /,
  overloaded to deal with all the allowed 3 permutations
  of matrices and doubles: (m,m), (m,d), (d,m).
Please don't kill me for that!

As you can see, the element wise mul, `amul`, is added.

Also, I changed the return types from void to CMatrix or CMatrix&.

I also removed the exception of division by zero, because:
(a) it's really not needed; the program doesn't crash, because
(b) floats support +/- inf, which is the expected result, which
(c) it's useful in computations in some contexts. Also,
(d) it makes the code more symmetric and uniform.

As a transitional step into a sane, uniform interface,
`getInverse()` is changed to return a value, rather than
modify its parameter.

As a transitional step into a better ordering of the functions
inside the cpp file, all operators are at the end of the file.

As a transitional step into completely separating the implementation
from the interface, each of `m.values[i][j]` is replaced by `m(i, j)`,
and each of `values[i][j]` or `this->values[i][j]` is replaced by
`(*this)(i, j)`.