refactor mercilessly
Removed most of the crap we got from the slides but actually not needed
in our project, including but not limited to most operators.

Removed many functions that are not actually needed, like getString().

Factored all the arithmatic functions into a macro and a few one-line functions.

Added a test case `make test` to the project, to detect if any change breaks it.
The test needs to be augemented more, though.

Reduced the precision, which is used to determine if two numbers are equal,
from 1e-15 to 1e-5. That helps with our use cases.

Made our project into a namespace, called `asu`.

Worked on some functions to ensure better interface.

Cleaned the headers a little bit.

Formated the exteremly short functions into one line or two,
not to waste the vertical space.