fix readCmd(); improve tests
readCmd now can understand a statement regardless how many
lines it's spread over, whether the user is a Windows user
or a GNU/Linux hacker. It does that while retaining the
ability to silent the output with a semicolon.

Uploaded the `bigexample.m` file.

Added a test file to test readCmd().