improve test cases' generating and Makefile
Following the refactoring efforts in `functions_test.cpp` by AbdoAshraf,
this commit updates `generate-functions-test-cases.m` to be compatible with
the new testing code. It also improves its formatting a little, and introduces
a workaround for a problem in the testing code, which caused every test
involving the last matrix in a file to fail.

As `functions_test.cpp` now reads the file generated by Octave directly,
`generate-functions-test-case.sh` and `mat2m.vim` are obsolete.
So, they are removed.

Also, in Makefile, `functions_test` is updated, and `readcmd_test` is removed.